iwlwifi: remove old intermediate 5.15+ firmwares
commit 3435843535f7 ("iwlwifi: remove old intermediate 5.15+ firmwares") upstream.

The iwlwifi firmware requirements since 5.15 for
for 20k series of chips was bumped from v66 to v72 for
AXxxx series in the 5.18 kernel so we can drop the
intermediate versions as they're now unused, it's
still the current highest rev firmware as of 6.2 cycle.

deepin-Intel-SIG: commit 3435843535f7 ("iwlwifi: remove old intermediate 5.15+ firmwares").

Cc: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Signed-off-by: Josh Boyer <jwboyer@kernel.org>
[ Furong Zhou: amend commit log ]
Signed-off-by: Furong Zhou <furong.zhou@intel.com>